+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## Unreleased
+## unreleased
+* Add `dataOnlyRequested` to `ThreeDSecureLookupRequest`
 * Add `chargeback_protection_level` into dispute search 
 
 ## 3.9.0